@@ -5,11 +5,11 @@ plugins {
 }
 
 group = 'com.lgcms'
-version = '0.0.1-SNAPSHOT'
+version = '0.1'
 
 java {
 	toolchain {
-		languageVersion = JavaLanguageVersion.of(17)
+		languageVersion = JavaLanguageVersion.of(21)
 	}
 }
 
@@ -35,8 +35,8 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.postgresql:postgresql'
+	implementation 'org.springframework.ai:spring-ai-starter-model-bedrock-converse'
 	implementation 'org.springframework.ai:spring-ai-starter-model-bedrock'
-	implementation 'org.springframework.ai:spring-ai-starter-vector-store-pgvector'
 	implementation 'org.springframework.boot:spring-boot-starter-validation'
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
