@@ -1,5 +1,7 @@
 package com.lgcms.leveltest.repository;
 
+import com.lgcms.leveltest.domain.Category;
+import com.lgcms.leveltest.domain.Difficulty;
 import com.lgcms.leveltest.domain.LevelTest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,12 +11,10 @@ import java.util.List;
 @Repository
 public interface LevelTestRepository extends JpaRepository<LevelTest, Long> {
 
-    // 카테고리와 난이도 기반 조회
-    List<LevelTest> findByCategoryAndDifficulty(String category, String difficulty);
+    List<LevelTest> findByCategoryAndDifficulty(Category category, Difficulty difficulty);
 
-    // 카테고리 기반 조회
-    List<LevelTest> findByCategory(String category);
+    List<LevelTest> findByCategory(Category category);
 
-    // 난이도 기반 조회
-    List<LevelTest> findByDifficulty(String difficulty);
+    List<LevelTest> findByDifficulty(Difficulty difficulty);
+
 }
